SCB-2176 Refactor rest router
