+++ pkg/rest/common.go
@@ -64,3 +64,31 @@ func isValidMethod(method string) bool {
 		return false
 	}
 }
+
+func match(s string, f func(c byte) bool, exclude byte, i int) (matched string, next byte, j int) {
+	j = i
+	for j < len(s) && f(s[j]) && s[j] != exclude {
+		j++
+	}
+
+	if j < len(s) {
+		next = s[j]
+	}
+	return s[i:j], next, j
+}
+
+func matchParticipial(c byte) bool {
+	return c != '/'
+}
+
+func isAlpha(ch byte) bool {
+	return ('a' <= ch && ch <= 'z') || ('A' <= ch && ch <= 'Z') || ch == '_'
+}
+
+func isDigit(ch byte) bool {
+	return '0' <= ch && ch <= '9'
+}
+
+func isAlnum(ch byte) bool {
+	return isAlpha(ch) || isDigit(ch)
+}
