+++ pkg/rest/handler.go
@@ -19,24 +19,42 @@ package rest
 
 import (
 	"net/http"
+	"net/url"
 )
 
-var (
-	serverHandler *ROAServerHandler // serverHandler is the default handler
-)
-
-func init() {
-	serverHandler = NewROAServerHander()
-	serverHandler.setChainName(ServerChainName)
+type urlPatternHandler struct {
+	Name string
+	Path string
+	http.Handler
 }
 
-// RegisterServant registers a ROAServantService into serverHandler
-// servant must be an pointer to service object
-func RegisterServant(servant interface{}) {
-	serverHandler.RegisterServant(servant)
-}
+func (roa *urlPatternHandler) try(path string) (p string, _ bool) {
+	var i, j int
+	l, sl := len(roa.Path), len(path)
+	for i < sl {
+		switch {
+		case j >= l:
+			if roa.Path != "/" && l > 0 && roa.Path[l-1] == '/' {
+				return p, true
+			}
+			return "", false
+		case roa.Path[j] == ':':
+			var val string
+			var nextc byte
+			o := j
+			_, nextc, j = match(roa.Path, isAlnum, 0, j+1)
+			val, _, i = match(path, matchParticipial, nextc, i)
 
-//GetRouter return the router fo REST service
-func GetRouter() http.Handler {
-	return serverHandler
+			p += url.QueryEscape(roa.Path[o:j]) + "=" + url.QueryEscape(val) + "&"
+		case path[i] == roa.Path[j]:
+			i++
+			j++
+		default:
+			return "", false
+		}
+	}
+	if j != l {
+		return "", false
+	}
+	return p, true
 }
